requirements.txt
Django==2.2.2
django-ckeditor==5.7.1
django-cors-headers==3.0.2
django-filter==2.1.0
django-js-asset==1.2.2
django-suit==2.0a1
djangorestframework==3.9.4
Pillow==6.0.0
pytz==2019.1
sqlparse==0.3.0
